pyproject.toml
[project]
name = "wechat-ai-assistant"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.77.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "wxauto>=3.9.11.17.5",
]
